Refactor tag and person package to category and transaction respectively
