+++ src/main/java/seedu/address/model/transaction/exceptions/TransactionNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person.exceptions;
+package seedu.address.model.transaction.exceptions;
 
 /**
  * Signals that the operation is unable to find the specified transaction.
