+++ src/main/java/seedu/address/storage/JsonAdaptedCategory.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.model.tag.Category;
+import seedu.address.model.category.Category;
 
 /**
  * Jackson-friendly version of {@link Category}.
