+++ src/main/java/seedu/address/ui/TransactionCard.java
@@ -7,7 +7,7 @@ import javafx.scene.control.Label;
 import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
-import seedu.address.model.person.Transaction;
+import seedu.address.model.transaction.Transaction;
 
 /**
  * An UI component that displays information of a {@code Transaction}.
