+++ src/test/java/seedu/address/logic/parser/AddressBookParserTest.java
@@ -22,8 +22,8 @@ import seedu.address.logic.commands.FindCommand;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.ListCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.NameContainsKeywordsPredicate;
-import seedu.address.model.person.Transaction;
+import seedu.address.model.transaction.NameContainsKeywordsPredicate;
+import seedu.address.model.transaction.Transaction;
 import seedu.address.testutil.EditTransactionDescriptorBuilder;
 import seedu.address.testutil.TransactionBuilder;
 import seedu.address.testutil.TransactionUtil;
