+++ src/test/java/seedu/address/model/category/CategoryTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model.tag;
+package seedu.address.model.category;
 
 import static seedu.address.testutil.Assert.assertThrows;
